issue #151 missed patches
git-svn-id: http://tradelink.googlecode.com/svn/trunk@989 ccc6a032-7b47-0410-9fd8-1f29c74ce1d2

+++ BrokerServers/TradeLibFast/TLServer_WM.h
@@ -22,7 +22,7 @@ namespace TradeLibFast
 		bool ENABLED;
 		__event void GotDebug(LPCTSTR msg);
 		CString debugbuffer;
-		static int TLSend(int type,LPCTSTR msg,CString windname);
+		static long TLSend(int type,LPCTSTR msg,CString windname);
 		void SrvGotOrder(TLOrder order);
 		void SrvGotFill(TLTrade trade);
 		void SrvGotTick(TLTick tick);
